tests: extract CommandOutput types to sub module
These types don't depend on TestEnvironment.
